# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartCheckout API",
        "description": "API for SmartCheckout demo (Laravel + Docker + JWT + queues).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "description": "Returns a paginated list of orders with optional filters.",
                "operationId": "6390e58970c6b576bd8c1ec198cb5fbb",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by order status (e.g. pending, paid, cancelled)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "paid"
                        }
                    },
                    {
                        "name": "customer_email",
                        "in": "query",
                        "description": "Filter by customer email",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "john@example.com"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of orders"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Checkout and create an order",
                "description": "Creates a new order from the given cart/customer data.",
                "operationId": "00de3580e94add04dce98ee82b2b2acd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer": {
                                        "description": "Customer data (existing or new)",
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "john@example.com"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "shipping_address": {
                                                "type": "string",
                                                "example": "Some street 1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "card"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "description": "Returns details of a single order.",
                "operationId": "29e4676537faf0c969a79eb39e944a39",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "description": "Cancels an order if it's in a cancellable state.",
                "operationId": "65b310314f60aa35656c74ebbf2abf8a",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully"
                    },
                    "422": {
                        "description": "Order cannot be cancelled (business rule violation)"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}/resync-wms": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Resync order to WMS",
                "description": "Dispatches a background job to resync the order to the WMS if it's eligible.",
                "operationId": "ce00658f94272c5ea24a69fbfea4ffdb",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "WMS sync dispatched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "WMS sync dispatched"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Order is not eligible for WMS sync (already synced or not paid)"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}